🏗️  consolidate helpers into notes
the test, remove, and find helpers are all note related and should
be accessible on the notes class
