+++ src/utils/notes.ts
@@ -2,6 +2,7 @@
 
 import fs from "fs"
 import path from "path"
+import { prompt } from "inquirer"
 import matter from "gray-matter"
 import { Config, ConfigurationKeys } from "."
 import { Frontmatter, FrontmatterKeys } from "./contentHelpers"
@@ -13,6 +14,37 @@ export class Notes extends Config {
         this.config = this.readConfig()
     }
 
+    /**
+     * Find a single note interactively
+     */
+    async find() {
+        const rootDir = this.config.get(ConfigurationKeys.NOTES_ROOT_DIR)
+        const questions = [
+            {
+                type: "fuzzypath",
+                name: "filePath",
+                excludePath: (nodePath: string) =>
+                    nodePath.startsWith("node_modules"),
+                excludeFilter: (nodePath: string) =>
+                    nodePath.startsWith(`${rootDir}/.`),
+                itemType: "file",
+                rootPath: rootDir,
+                message:
+                    "Select the note you'd like to update (excludes .dotfiles):",
+                default: "",
+                suggestOnly: false,
+                depthLimit: 0,
+            },
+        ]
+
+        return await prompt(questions).then(
+            (answers) => answers && (answers.filePath as string)
+        )
+    }
+
+    /**
+     * List all notes that are managed by Nom
+     */
     list() {
         const rootDir = this.readConfig().get(ConfigurationKeys.NOTES_ROOT_DIR)
         if (!rootDir) {
@@ -29,16 +61,31 @@ export class Notes extends Config {
     }
 
     /**
-     *
+     * Read a single note given a filepath
      * @param filePath
      * @return {} - An object of Frontmatter & the file path
      */
-    readNote(filePath: string) {
+    read(filePath: string) {
         return matter.read(filePath)
     }
 
+    remove(filePath: string) {
+        // export function removeNoteFile(filePath: string) {
+        if (!this.testPath(filePath)) {
+            throw new Error(`No file exists at path ${filePath}`)
+        }
+        fs.unlinkSync(filePath)
+    }
+
+    async testPath(filePath: string) {
+        return await fs.promises
+            .access(filePath)
+            .then(() => true)
+            .catch(() => false)
+    }
+
     /**
-     *
+     * Get the frontmatter for a single note
      * @param filePath
      * @return {} - An object of Frontmatter & the file path
      */
@@ -48,13 +95,17 @@ export class Notes extends Config {
         }
         return {
             path: filePath,
-            ...this.readNote(filePath).data,
+            ...this.read(filePath).data,
         } as Frontmatter & { path: string }
     }
 
+    /**
+     * Get the frontmatter for all notes managed by Nom,
+     * Notes where there is only a key are filtered out (e.g., `.contents` which has no frontmatter).
+     */
     async allNotesFrontmatter() {
         return (await this.list())
             .map((file) => this.getFrontmatter(file))
-            .filter((note) => Object.keys(note).length > 1) // filter out any notes where the only key is path
+            .filter((note) => Object.keys(note).length > 1)
     }
 }
